% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions_v6.r
\name{analyse.image}
\alias{analyse.image}
\title{Analyse an image or a set of images}
\usage{
analyse.image(path.sample, path.result, path.image, file.image = NA,
  leaf.area.min, leaf.border, lesion.border, lesion.area.min,
  lesion.area.max, lesion.eccentricity.max, lesion.color, blur.diameter)
}
\arguments{
\item{path.sample}{The path of the directory containing the sampled images used for training. After the training step, this directory contains the parameters of the training set.}

\item{path.result}{The path of the directory where to store the result files (created by the function if it does not exist).}

\item{path.image}{The path of the directory containing the images to analyse.}

\item{file.image}{A character vector containg the fils names of the images to analyse in path.image (NA to analyse all the images in path.image).}

\item{leaf.area.min}{The minimum area of a leaf (in pixels).}

\item{leaf.border}{The diameter of the brush (in pixels) used to erode the leaf border.}

\item{lesion.border}{The diameter of the brush (in pixels) used to erode the lesion border.}

\item{lesion.area.min}{The minimum area of a lesion (in pixels).}

\item{lesion.area.max}{The maximum area of a lesion (in pixels).}

\item{lesion.eccentricity.max}{The maximum eccentricity of a lesion.}

\item{lesion.color}{Lesion color in the output image.}

\item{blur.diameter}{The diameter of the brush (in pixels) used to blur the image (0 for no blur).}
}
\description{
Analyse an image or a set of images
}
