% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions_v6.r
\name{analyseImages}
\alias{analyseImages}
\title{Analyse an image or a set of images}
\usage{
analyseImages(pathTraining, pathResult, pathImages = NULL,
  fileImage = NULL, leafAreaMin = 1000, leafBorder = 5,
  lesionBorder = 3, lesionAreaMin = 10, lesionAreaMax = 120000,
  lesionEccentricityMin = 0, lesionEccentricityMax = 1,
  lesionColorBorder = "#0000FF11", lesionColorBodies = "#FE8E0000",
  blurDiameter = 0, outPosition = "right", parallelThreadsNum = 1,
  mode = "CMD")
}
\arguments{
\item{pathTraining}{The path of the directory containing the sampled images used for training. After the training step, this directory contains the parameters of the training set.}

\item{pathResult}{The path of the directory where to store the result files (created by the function if it does not exist).}

\item{pathImages}{The path of the directory containing the images to analyse.}

\item{fileImage}{A character vector containg the fils names of the images to analyse in pathImages (NULL to analyse all the images in pathImages).}

\item{leafAreaMin}{The minimum area of a leaf (in pixels) Default:1000.}

\item{leafBorder}{The diameter of the brush (in pixels) used to erode the leafBorder Default:5.}

\item{lesionBorder}{The diameter of the brush (in pixels) used to erode the lesionBorder Default:3.}

\item{lesionAreaMin}{The minimum area of a lesion (in pixels) Default:10.}

\item{lesionAreaMax}{The maximum area of a lesion (in pixels) Default:120000.}

\item{lesionEccentricityMin}{The minimum eccentricity of a lesion Default:0.}

\item{lesionEccentricityMax}{The maximum eccentricity of a lesion Default:1.}

\item{lesionColorBorder}{hexadecimal code for output fill color for lesion in the output image Default:#0000FF (blue).}

\item{lesionColorBodies}{hexadecimal code for output bodies color for lesion in the output image Default:#FE8E0000 (transparent).}

\item{blurDiameter}{The diameter of the brush (in pixels) used to blur the image (0 for no blur) Default:0)'.}

\item{outPosition}{join origale and color lesion image at right or buttom Default:right)'.}

\item{parallelThreadsNum}{number of thread use, 1 thread analysis 1 image if >= 2 Default:1)'.}

\item{mode}{auto selection to switch between GUI or CMD mode Default:"CMD")'.}
}
\description{
Analysis step can use many ram on parallel mode.
The function return a dataframe with file name, exit status and message if error.
}
\examples{
dataframeExitStatus <- analyseImages(pathTraining = "../exemple1/learning",
pathResult = "../exemple1/results",
pathImages = "../exemple1/samples", parallelThreadsNum=4)
}
