% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{resizeImageDirectory}
\alias{resizeImageDirectory}
\title{Resize all images on directory in order to reduce size}
\usage{
resizeImageDirectory(path, factor = 2, mode = "CMD")
}
\arguments{
\item{path}{The path of the directory containing the sampled images to resize.}

\item{factor}{The factor of reduce size Default: 2.}
}
\description{
The function build new folder with images resize.
}
\examples{
pathImages <- '../exemple1/samples'
resizeImageDirectory(pathImages, 1.5)
}
