% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{splitImages}
\alias{splitImages}
\title{Split all images on directory}
\usage{
splitImages(path, splitVertical = 1, splitHorizontal = 1,
  numOrder = "bottum", mode = "CMD")
}
\arguments{
\item{path}{The path of the directory containing the sampled images to split.}

\item{splitVertical}{The number of part split vertical Default: 1.}

\item{splitHorizontal}{The number of part split horizontal Default: 1.}

\item{numOrder}{The order to numerote output image left from right or top to buttom. Default:bottum).}
}
\description{
The function split image on n horizontal / m vertical.
For exemple if you want to split on 2 part
}
\examples{
pathImages <- '../exemple1/samples'
splitImages(pathImages, splitVertical = 2, splitHorizontal = 3) # split on 6 part (2x3)
}
