% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{splitImages}
\alias{splitImages}
\title{Split all images on directory}
\usage{
splitImages(path, splitVertical = 1, splitHorizontal = 1,
  marginTop = 0, marginRight = 0, marginBottom = 0, marginLeft = 0,
  numOrder = "bottum", mode = "CMD")
}
\arguments{
\item{path}{The path of the directory containing the sampled images to split.}

\item{splitVertical}{The number of part split vertical Default: 1.}

\item{splitHorizontal}{The number of part split horizontal Default: 1.}

\item{marginTop}{The crop marge top. Default:0}

\item{marginRight}{The crop marge right. Default:0}

\item{marginBottom}{The crop marge bottom. Default:0}

\item{marginLeft}{The crop marge left. Default:0}

\item{numOrder}{The order to numerote output image left from right or top to buttom. Default:bottum.}
}
\description{
The function split image on n horizontal / m vertical.
For exemple if you want to split on 2 part
}
\examples{
pathImages <- '../exemple1/samples'
splitImages(pathImages, splitVertical = 2, splitHorizontal = 3, marginTop = 10, marginRight = 300, marginBottom = 300, marginLeft = 170,) # split on 6 part (2x3)
}
