% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_apprentissage_V5.r
\name{training}
\alias{training}
\title{Launch training function to build set}
\usage{
training(path.sample)
}
\arguments{
\item{path.sample}{Path to folder with image for calibration}
}
\value{

}
\description{
Build the training on calibration folders.
}
\details{
This function ...
Training input folder must include sub-folders:
\itemize{
  \item limb
  \item background
  \item lesion
}
}
\examples{
path.sample <- './Exemple2_Stella/Samples'
training(path.sample)
}
