% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_functions_V6.r
\name{training}
\alias{training}
\title{Compute and saves on disk the parameters of the training set}
\usage{
training(pathTraining, method = "lda", transform = NULL,
  colormodel = "rgb", mode = "CMD")
}
\arguments{
\item{pathTraining}{The path of the folder containing sampled images for training. This folder must contain at least 3 sub-folders with name 'background', 'limb' and 'lesion'.}

\item{method}{Method of discrimainant analysis: "lda" (default) or "qda"}

\item{transform}{Function for data transformation before analysis (e.g. sqrt)}

\item{colormodel}{Model of color for the analysis: "rgb" (default) or "hsv"}

\item{mode}{auto selection to switch between GUI or CMD mode Default:"CMD")'.}
}
\description{
Training input folder must include sub-folders:
\itemize{
  \item limb
  \item background
  \item lesion
}
This sub-folder can contain either image files or sub-folders containing different groups of image files
The function return the confusion matrix and error rate.
}
\examples{
pathTraining <- '../Exemple1/learning/' ## FOR all OS (linux Mac Windows)
pathTraining <- '..\\\\Exemple1\\\\learning' ## FOR windows only
confusionMatrix <- training(pathTraining)
confusionMatrix <- training(pathTraining, transform=function(x) log1p(x),colormodel='rgb', method='svm')
confusionMatrix <- training(pathTraining, colormodel='hsv', method='lda')
confusionMatrix <- training(pathTraining, transform=function(x) (sin(pi*(x-0.5))+1)/2, method='qda')
confusionMatrix <- training(pathTraining, transform=function(x) asin(2*x-1)/pi+0.5)
confusionMatrix <- training(pathTraining, transform=log1p)
}
