% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_functions_V6.r
\name{training}
\alias{training}
\title{Compute and saves on disk the parameters of the training set}
\usage{
training(path.training, method = "lda", transform = NULL,
  colormodel = "rgb")
}
\arguments{
\item{path.training}{The path of the folder containing sampled images for training. This folder must contain at least 3 sub-folders with name 'background', 'limb' and 'lesion'.}

\item{method}{Method of discrimainant analysis: "lda" (default) or "qda"}

\item{transform}{Function for data transformation before analysis (e.g. sqrt)}

\item{colormodel}{Model of color for the analysis: "rgb" (default) or "hsv"}
}
\description{
Training input folder must include sub-folders:
\itemize{
  \item limb
  \item background
  \item lesion
}
This sub-folder can contain either image files or sub-folders containing different groups of image files
}
\examples{
path.training <- '../Exemple_Dominique/CLFD/Samples/Sup'
training(path.training)
training(path.training, transform=function(x) log1p(x),colormodel='rgb', method='svm')
training(path.training, colormodel='hsv', method='lda')
training(path.training, transform=function(x) (sin(pi*(x-0.5))+1)/2, method='qda')
training(path.training, transform=function(x) asin(2*x-1)/pi+0.5)
training(path.training, transform=log1p)
}
